requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.0.1
gunicorn==23.0.0
python-dotenv==1.0.0
requests==2.32.3
requests-mock==1.12.1
Authlib==1.2.1
mysqlclient==2.2.4
selenium==4.27.1
pytest==8.3.3
pytest-mock==3.14.0
locust==2.32.4
